--- src/components/envvarwidget.tsx
@@ -1,14 +1,0 @@
-//import React, {useState} from 'react';
-
-// export const EnvVariablesField = (props: any) => {
-// 	return (
-//     <div id="custom">
-//       <p>Yeah, I'm pretty dumb.</p>
-//       <div>My props are: {JSON.stringify(props)}</div>
-//     </div>
-//   );
-// };
-
-// export const envvarsfield = {
-//   SchemaField: EnvVariablesField
-// };
