+++ src/components/fields/ipymenu.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 
 /*
  * This is a nested comoponenet in the FieldTemplate, and
@@ -23,79 +23,6 @@ export const IPythonFormGroup = (props: {
           {props.properties[index].content}
         </div>
       ))}
-    </div>
-  );
-};
-
-export const KeyValueWidget = (props: any) => {
-  const [key, setKey] = useState(props.value.name);
-  const [val, setVal] = useState(props.formData[props.value.name]);
-  return (
-    <div>
-      <input
-        type="string"
-        value={key}
-        onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
-          props.handleKey(key, e.target.value);
-          setKey(e.target.value);
-        }}
-      />
-      <input
-        type="string"
-        value={val}
-        onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
-          props.handleVal(key, e.target.value);
-          setVal(e.target.value);
-        }}
-      />
-    </div>
-  );
-};
-
-/*
- * This function is a field
- * that is rendered for the environment variable.
- */
-export const EnvVarForm = (props: any) => {
-  /*
-   * Function to handle a change of key values in a key value store
-   */
-  const handleKeyChange = (key: string, newkey: string) => {
-    props.formData[newkey] = props.formData[key];
-    delete props.formData[key];
-    console.log(props.formData);
-  };
-
-  /*
-   * Function to handle any changing values
-   */
-  const handleValueChange = (key: string, newvalue: string) => {
-    props.formData[key] = newvalue;
-    console.log(props.formData);
-  };
-
-  const widget: (props: any) => JSX.Element = props.uiSchema["ui:widget"];
-  const formData = props.formData;
-  return (
-    <div>
-      {props.properties.map((item: any) =>
-        widget(
-          (props = {
-            value: item,
-            formData: formData,
-            handleKey: handleKeyChange,
-            handleVal: handleValueChange,
-          })
-        )
-      )}
-      <button
-        type="button"
-        onClick={(e: any) => {
-          alert(props.onAddClick);
-        }}
-      >
-        Add New
-      </button>
     </div>
   );
 };
